pyproject.toml
[tool.poetry]
name = "audiokit"
version = "0.1.0"
description = "Audio processing toolkit with AI-driven features"
authors = ["Nathaniel Houk <nate.houk@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/epsilon-records/audiokit"
repository = "https://github.com/epsilon-records/audiokit"
documentation = "https://github.com/epsilon-records/audiokit/docs"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
loguru = "^0.7.2"
torch = "^2.2.0"
torchaudio = "^2.2.0"
numpy = "^1.26.0"
llama-index = "0.12.15"
llama-index-vector-stores-pinecone = "0.4.4"
openai = "^1.12.0"
typer = "^0.9.0"
rich = "^13.7.0"
soundfile = "^0.12.1"
python-dotenv = "^1.0.0"
llama-index-llms-openai-like = "0.3.3"
llama-index-embeddings-huggingface = "0.5.1"

[tool.poetry.group.dev.dependencies]

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
ruff = "^0.2.2"
soundfile = "^0.12.1"
numpy = "^1.26.4"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.9"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-x --cov=audiokit --cov-report=term-missing"
testpaths = [
    "tests",
]
log_cli = true
log_level = "INFO"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.poetry.scripts]
audiokit = "audiokit.cli:app"